+++ source/standalone_examples/api/omni.isaac.ros2_bridge/rtx_lidar.py
@@ -55,7 +55,7 @@ _, sensor = omni.kit.commands.execute(
     parent=None,
     config="Example_Rotary",
     translation=(0, 0, 1.0),
-    orientation=Gf.Quatd(0.5, 0.5, -0.5, -0.5),  # Gf.Quatd is w,i,j,k
+    orientation=Gf.Quatd(1.0, 0.0, 0.0, 0.0),  # Gf.Quatd is w,i,j,k
 )
 
 # RTX sensors are cameras and must be assigned to their own render product